extract from template to base skill
