+++ ovos_utils/skills/templates/media_collection.py
@@ -2,9 +2,7 @@ from ovos_utils.waiting_for_mycroft.common_play import CommonPlaySkill, \
     CPSMatchLevel, CPSTrackStatus, CPSMatchType
 from os.path import join, dirname, basename
 import random
-import re
-import datetime
-from ovos_utils import get_mycroft_root
+from ovos_utils import get_mycroft_root, datestr2ts
 from ovos_utils.log import LOG
 from ovos_utils.parse import fuzzy_match
 from json_database import JsonStorageXDG
@@ -26,14 +24,6 @@ try:
 except ImportError:
     LOG.error("py_VOD not installed!")
     LOG.debug("py_VOD>=0.3.1")
-
-
-def datestr2ts(datestr):
-    y = int(datestr[:4])
-    m = int(datestr[4:6])
-    d = int(datestr[-2:])
-    dt = datetime.datetime(y, m, d)
-    return dt.timestamp()
 
 
 class MediaCollectionSkill(CommonPlaySkill):
@@ -237,22 +227,6 @@ class MediaCollectionSkill(CommonPlaySkill):
                 best_score = score
                 leftover_text = phrase.replace(title, "")
         return match, best_score, best_video, leftover_text
-
-    def remove_voc(self, utt, voc_filename, lang=None):
-        lang = lang or self.lang
-        cache_key = lang + voc_filename
-
-        if cache_key not in self.voc_match_cache:
-            # trigger caching
-            self.voc_match(utt, voc_filename, lang)
-
-        if utt:
-            # Check for matches against complete words
-            for i in self.voc_match_cache[cache_key]:
-                # Substitute only whole words matching the token
-                utt = re.sub(r'\b' + i + r"\b", "", utt)
-
-        return utt
 
     def normalize_title(self, title):
         title = title.lower().strip()
